+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * BREAKING CHANGE: The `m` parameter in in `Flay` and `Flayable` has been
   existentialized, to be chosen by the caller.
 
-* Added `Flayable1`, `flay1`, `trivial`, `collect1`.
+* Added `Flayable1`, `flay1`, `trivial`, `collect1`, `Record`, `Unit`, `unit`.
 
 
 # Version 0.1